requirements.txt
python-telegram-bot==20.3
SQLAlchemy==2.0.20
python-dotenv==1.0.0
psycopg2-binary==2.9.9
